nikkei_project/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-auth==1.26.1
google-auth-oauthlib==0.4.2
gspread==3.6.0
httplib2==0.19.0
idna==2.10
numpy==1.20.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.3
